。。。

+++ process/manager/process.go
@@ -3,6 +3,7 @@ package manager
 import (
 	"context"
 	"fmt"
+	"github.com/gogf/gf/v2/container/gmap"
 	"github.com/gogf/gf/v2/errors/gerror"
 	"github.com/gogf/gf/v2/os/glog"
 	"github.com/gogf/gf/v2/os/gproc"
@@ -58,53 +59,76 @@ func get(os string, ctx context.Context, isDebug bool, logger *glog.Logger) Mana
 
 type operation struct {
 	manager *gproc.Manager
+	mapping *gmap.StrIntMap
 	err     error
 }
 
-func (o *operation) GetProcess(pid int) *gproc.Process {
+func (o *operation) GetProcess(name string) *gproc.Process {
+	pid := o.GetPid(name)
+	if pid < 1 {
+		return nil
+	}
 	return o.manager.GetProcess(pid)
 }
 
-func (o *operation) GetPid(name string) *gproc.Process {
-	return nil
+func (o *operation) GetPid(name string) int {
+	if o.err != nil {
+		logger.Errorf(context.Background(), "get process %s failed ,error: %s", name, o.err.Error())
+		return -1
+	}
+	if len(name) < 1 {
+		return -1
+	}
+	return o.mapping.Get(name)
 }
 
 // Start 启动服务
 func (o *operation) Start(ctx context.Context, runtime *RuntimeConfig, logger *glog.Logger) (int, error) {
-	//process := o.manager.NewProcess(runtime.Path, runtime.Cmd, runtime.Env)
 	process := gproc.NewProcessCmd(gstr.Join(runtime.Cmd, " "), runtime.Env)
 	process.Manager = o.manager
 	pid, err := process.Start(ctx)
-	runtime.Pid = pid
 	if err != nil {
-		return runtime.Pid, gerror.NewCode(consts.FAILURE, fmt.Sprintf("\n -- [ %s ] 进程启动失败: %s", runtime.Name, err.Error()))
+		return pid, gerror.NewCode(consts.FAILURE, fmt.Sprintf("\n -- [ %s ] 进程启动失败: %s", runtime.Name, err.Error()))
 	}
 	go func() {
 		defer func() {
 			if err := process.Wait(); err != nil {
-				logger.Errorf(ctx, "\n -- [ %s ] 进程停止失败 。。。 [ PID ：%d ][ Error ：%s ]", runtime.Name, runtime.Pid, err.Error())
+				logger.Errorf(ctx, "\n -- [ %s ] 进程启动失败 。。。 [ PID ：%d ][ Error ：%s ]", runtime.Name, pid, err.Error())
 			}
 		}()
 	}()
-	logger.Infof(ctx, "\n -- [ %s ] 进程启动成功 。。。 [ PID ：%d ]", runtime.Name, runtime.Pid)
-	return runtime.Pid, nil
+	o.mapping.Set(runtime.Name, pid)
+	logger.Infof(ctx, "\n -- [ %s ] 进程启动命令提交成功 。。。 [ PID ：%d ]", runtime.Name, pid)
+	return pid, nil
 }
 
 // Restart 重启服务
 func (o *operation) Restart(ctx context.Context, runtime *RuntimeConfig, logger *glog.Logger) (int, error) {
-	process := o.manager.GetProcess(runtime.Pid)
+	if o.err != nil {
+		return -1, o.err
+	}
+	pname := runtime.Name
+	if len(pname) < 1 {
+		return -1, gerror.NewCode(consts.FAILURE, "\n -- 进程重启失败，进程名称不能为空")
+	}
+	pid := o.GetPid(runtime.Name)
+	process := o.manager.GetProcess(pid)
 	if process != nil {
 		err := process.Kill()
 		if err != nil {
-			logger.Errorf(ctx, "\n -- [ %s ] 进程停止失败 。。。 [ PID ：%d ][ Error ：%s ]", runtime.Name, runtime.Pid, err.Error())
+			logger.Errorf(ctx, "\n -- [ %s ] 进程停止失败 。。。 [ PID ：%d ][ Error ：%s ]", runtime.Name, pid, err.Error())
 		}
 	}
 	return o.Start(ctx, runtime, logger)
 }
 
 // Stop 停止服务
-func (o *operation) Stop(ctx context.Context, runtime *RuntimeConfig, logger *glog.Logger) error {
-	process := o.manager.GetProcess(runtime.Pid)
+func (o *operation) Stop(ctx context.Context, name string, logger *glog.Logger) error {
+	if o.err != nil {
+		return o.err
+	}
+	pid := o.GetPid(name)
+	process := o.manager.GetProcess(pid)
 	if process == nil {
 		return nil
 	}
@@ -112,13 +136,18 @@ func (o *operation) Stop(ctx context.Context, runtime *RuntimeConfig, logger *gl
 	if err == nil {
 		return nil
 	}
-	logger.Errorf(ctx, "\n -- [ %s ] 进程停止失败 。。。 [ PID ：%d [ Error ：%s ]", runtime.Name, runtime.Pid, err.Error())
+	o.mapping.Remove(name)
+	logger.Errorf(ctx, "\n -- [ %s ] 进程停止失败 。。。 [ PID ：%d [ Error ：%s ]", name, pid, err.Error())
 	return err
 }
 
 // Status 服务状态
-func (o *operation) Status(ctx context.Context, runtime *RuntimeConfig) health.Health {
-	process := o.manager.GetProcess(runtime.Pid)
+func (o *operation) Status(name string) health.Health {
+	if o.err != nil {
+		return health.UNKNOWN
+	}
+	pid := o.GetPid(name)
+	process := o.manager.GetProcess(pid)
 	if process == nil {
 		return health.UNKNOWN
 	}
@@ -130,27 +159,33 @@ func (o *operation) Status(ctx context.Context, runtime *RuntimeConfig) health.H
 }
 
 func (o *operation) Clear() {
-	o.manager.KillAll()
+	if o.err != nil {
+		return
+	}
+	if err := o.manager.KillAll(); err == nil {
+		o.mapping.Clear()
+	}
 }
 
 // init
 func (o *operation) init(ctx context.Context, isDebug bool, logger *glog.Logger) {
 	o.manager = gproc.NewManager()
+	o.mapping = gmap.NewStrIntMap(true)
 }
 
 type Manager interface {
 	// GetProcess 获取进程对象
-	GetProcess(pid int) *gproc.Process
+	GetProcess(name string) *gproc.Process
 	// GetPid 获取进程ID
-	GetPid(name string) *gproc.Process
+	GetPid(name string) int
 	// Start 启动服务
 	Start(ctx context.Context, runtime *RuntimeConfig, logger *glog.Logger) (int, error)
 	// Restart 重启服务
 	Restart(ctx context.Context, runtime *RuntimeConfig, logger *glog.Logger) (int, error)
 	// Stop 停止服务
-	Stop(ctx context.Context, runtime *RuntimeConfig, logger *glog.Logger) error
+	Stop(ctx context.Context, name string, logger *glog.Logger) error
 	// Status 状态服务查询
-	Status(ctx context.Context, runtime *RuntimeConfig) health.Health
+	Status(name string) health.Health
 	// Clear 清空进程信息
 	Clear()
 	// 初始化
@@ -158,7 +193,6 @@ type Manager interface {
 }
 
 type RuntimeConfig struct {
-	Pid     int      `json:"pid,omitempty"`
 	Name    string   `json:"name,omitempty"`
 	Restart string   `json:"restart,omitempty"`
 	Cmd     []string `json:"cmd,omitempty"`
